mv util func to a new file
